Moved GraphFilterPanel MVC to graph_filter.py
